Remove T4 rumble LED
This is a nice feature to have in theory, but because the built-in LED is also one of the side pins this removes one of the available inputs with no options to override it.

To that end it might be worth adding as a board option to the IDE later on. In the meantime, this can be implemented by the user with the library callback if need be.